requirements.txt
openai==1.7.2
python-dotenv==1.0.0
streamlit==1.30.0
